+++ src/battle/battleConfig.js
