+++ src/battle/data/effectRegistry.js
