+++ src/battle/data/effects.js
