+++ src/embeds/pokemonEmbeds.js
@@ -15,8 +15,8 @@ const {
   typeConfig,
   growthRateConfig,
 } = require("../config/pokemonConfig");
-const { abilityConfig } = require("../battleEngine/battleConfig");
-const { getMove } = require("../battleEngine/data/moveService");
+const { abilityConfig } = require("../battle/battleConfig");
+const { getMove } = require("../battle/data/moveService");
 const {
   getWhitespace,
   getPBar,
