+++ src/services/spawn.js
@@ -3,7 +3,7 @@
 const { ButtonStyle } = require("discord.js");
 const { buildButtonActionRow } = require("../components/buttonActionRow");
 const { backpackItems } = require("../config/backpackConfig");
-const { typeAdvantages } = require("../battleEngine/battleConfig");
+const { typeAdvantages } = require("../battle/battleConfig");
 const { eventNames } = require("../config/eventConfig");
 const { pokeballConfig } = require("../config/gachaConfig");
 const {
